[FE] move: BaseCard 컴포넌트 common 하위로 이동 #115
