+++ frontend/src/components/Cards/CartProductCard.vue
@@ -19,7 +19,7 @@
 
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import BaseCard from '@components/Cards/BaseCard.vue';
+import BaseCard from '@components/Cards/common/BaseCard.vue';
 import Icon from '@components/Common/Icon.vue';
 
 @Component({
